* commit '7a9ea2fd853bd65c2554d0c716f458b1fd485d0b':
  APM-166915 get rid of some compiler warnings

GitOrigin-RevId: 7b2a46b322fcd37f553c9ef86d19079d9a084e2c

+++ src/core/communication/BeaconSendingContext.cxx
@@ -290,7 +290,7 @@ std::vector<std::shared_ptr<core::objects::SessionInternals>> BeaconSendingConte
 	return finishedSessions;
 }
 
-int32_t BeaconSendingContext::getSessionCount()
+size_t BeaconSendingContext::getSessionCount()
 {
 	return mSessions.size();
 }
